Team_69/aibackend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
numpy
Pillow
tensorflow
pymongo
python-jose
python-multipart

